Allow memory storage to handle subscriptions async only optionally
By this the default behavior is to wait for subscriptions to handle a certain event,
then resume the processing in the code which triggered an event.
This should increase the execution speed of tests where subscriptions used for state management.

Optionally the async subscription event handling can be still turned on,
and then the Creator, Updater and Deleter action will not hang until the subscribers are done.