Unify memory storage options by moving event log disabling to storage options
This approach should make it easier to set up memory storage for a given purpose.
Also, all options default to false and act as a flag that can be turned on,
instead of representing states.

This design of the options should allow a more natural usage of the storage,
without a need for using factory function to initialize a new instance with a reasonable default value set.