align package naming with TDD practices
The package name "specs" not that well used in the TDD terminologies to describe a contract between a consumer and a supplier.
The eiffel language established the "contract" word usage, as well the "design by contract" software correctness methodology.
Since the package goal to support this type of design approaches, I find contracts a better suiting package name.