remove generics extension enforcement from rfc7807 error handler
For the majority of the cases, it does not add additional value to enforce generics in the DTO,
so removing the generic type parameter makes the error hadler more reusable as an error DTO generator.

For backward compatibility support, an DTOExt type was also supplied which still accepts the type parameter.