introducing stringcase package
The stringcase package makes it simple to change the style of strings between formats like snake_case or PascalCase. This is handy when you want to maintain consistent style among map string keys that people need to read.

For instance, when you create logs, you may accidentally use different styles for keys. However, fixing these later may not be possible due to constraints in the logging system. As a precaution, you can format all the logging field keys to a particular style using the package, so that you are sure they all look the same.