rename "metrics" to "details" in /health check endpoint based on community feedback
If the field is called as "metrics",
some people might mistakenly assume
that the health check endpoint details replace application metrics,
but this is not the case.

Tools like Prometheus are better suited for metrics
and have strict rules on what format the exported metric should be.

Health check endpoint details provide insight into what happened along with an error,
to help an investigation, rather than to help an automated application monitoring system.