decouple logging and the logger package
The new decoupled logger package clarifies its intended use.
This logger package serves as a global application log entity,
also making it convenient for testing purposes.
It is configurable to use any logger or logging output needed.
It streamlines all frameless packages to use the same logging package
without the need to inject a logger everywhere.

The logging package on the other hand gives a backbone to logging related requirements.