introduce resilience package
This package brings together all the components
related to resilience enhancement into one place,
as they share the same bounded context.

As we continue to develop more implementations focused on improving resilience,
having them in a single location makes them more compatible
and easier to work with in a unified way.

Additionally, this change reduces the number of packages in `framless/pkg`,
which currently can be overwhelming for those new to the project.

We’ve also addressed a small but notable inconvenience:
when typing "return", the auto-completion no longer defaults to import "retry".