BREAKING CHANGE: retire port/iterators and replace it with the stdlib iter package
As part of our ongoing effort to reduce projects being dependent on frameless, we are retiring the iterators port.
Until now, the iterators port has been an outlier since it wasn’t a type that could be replaced directly from the standard library.

With the introduction of the standard library’s iter package,
it is now possible to define a crud query many method signature using the builtin iterator.

While this is a breaking change, it simplifies development.
Developers will need to learn even less about Frameless to follow its conventions—and in many cases,
they won’t even need to be aware that they are working within Frameless conventions when using CRUD operations.