make enum registration safer against developer mistakes
When defining enums in a hurry, a developer might forget to set the type as const.
As a result, the enum values default to a built-in type, which is often incorrect.
This can lead to unexpected validation errors in parts of the code where the developer wouldn’t anticipate them.

To avoid mitigate this, `enum.Register` will panic when a registration is requested for a Go built-in type.