release
Added support for the `Validate() error` method in the validate and env packages.

Introduced a `SingleUse` type in iterkit for documentation purposes,
making it easier to indicate at the type level whether an iterator is intended for single use.

Add safety mechanism to avoid unintentional enum registration for a built-in type.