introduce LockstepReaders in iokit
LockstepReader allows reading the same io.Reader multiple times on concurrent goroutines.
The benefit of doing so is that if we want to use utility functions which utilise the io.Reader API,
to achieve their functionality,
But we would need to do this multiple times for different purposes,
but we wish to avoid reading the whole content into memory,
then we could do this using the iokit.NewLockstepReaders utility.

One example usage for this would be where we want to consume a json stream input,
and for different lexical parts of the data structure, we want to use a different json decoder,
then we can reuse the same input source thourgh lockstep phased mirrored reading between the json decoders.